clean up Project API

+++ backend/src/main/java/com/felipesouza/portfolio/project/ProjectService.java
@@ -1,12 +1,14 @@
 package com.felipesouza.portfolio.project;
 
 
+import com.felipesouza.exceptions.InvalidUuidException;
 import com.felipesouza.exceptions.ProjectNotFoundException;
 import com.felipesouza.util.DateHandler;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 
@@ -14,23 +16,13 @@ import java.util.stream.Collectors;
 public class ProjectService {
     private final ProjectRepository repository;
     private final ProjectDTOMapper dtoMapper;
+    private static final Pattern UUID_V4_REGEX = Pattern.compile(
+            "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
+    );
 
     public ProjectService(ProjectRepository repository, ProjectDTOMapper dtoMapper) {
         this.repository = repository;
         this.dtoMapper = dtoMapper;
-    }
-
-
-    /**
-     * Finds a project based on it's namespace and returns it's ID.
-     * @param namespace The ProjectEntity's namespace.
-     * @return The ProjectEntity's ID (UUIDv4).
-     * @throws ProjectNotFoundException When the namespace doesn't belong to any ProjectEntity.
-     */
-    private String getProjectId(String namespace) throws ProjectNotFoundException {
-        ProjectEntity projectEntity = repository.findByNamespace(namespace)
-                .orElseThrow(() -> new ProjectNotFoundException("Namespace doesn't belong to any ProjectEntity"));
-        return projectEntity.getId();
     }
 
     /**
@@ -39,7 +31,11 @@ public class ProjectService {
      * @return The DTO of the ProjectEntity.
      * @throws ProjectNotFoundException When the ID doesn't belong to any ProjectEntity.
      */
-    public ProjectDTO getProjectById(String projectId) throws ProjectNotFoundException {
+    public ProjectDTO getProjectById(String projectId) throws ProjectNotFoundException, InvalidUuidException {
+        if (isUUIDInvalid(projectId)) {
+            throw new InvalidUuidException();
+        }
+
         ProjectEntity project = this.repository.findById(projectId)
                 .orElseThrow(() -> new ProjectNotFoundException("ID not found"));
         return dtoMapper.apply(project);
@@ -67,33 +63,32 @@ public class ProjectService {
                 .collect(Collectors.toList());
     }
 
-    public List<SimpleProjectDTO> getSimplifiedProjects() {
-        return getProjects()
-                .stream()
-                .map(ProjectDTOMapper::simplify)
-                .collect(Collectors.toList());
-    }
-
 
     /**
      * Deletes the ProjectEntity with given ID.
      * @param projectId The ProjectEntity's ID (UUIDv4).
      */
     public void deleteProjectById(String projectId) {
+        if (isUUIDInvalid(projectId)) {
+            throw new InvalidUuidException();
+        }
         this.repository.deleteById(projectId);
     }
 
 
     /**
      * Updates the ProjectEntity with the given ID.
-     * @param id The ProjectEntity's ID (UUIDv4).
+     * @param projectId The ProjectEntity's ID (UUIDv4).
      * @param request The request with the values to be edited.
      * @throws ProjectNotFoundException When the ID doesn't belong to any ProjectEntity.
      */
-    public void updateProjectById(String id, ProjectAddRequest request) throws ProjectNotFoundException
-    {
+    public void updateProjectById(String projectId, ProjectAddRequest request) throws ProjectNotFoundException  {
+        if (isUUIDInvalid(projectId)) {
+            throw new InvalidUuidException();
+        }
+
         try {
-            ProjectDTO oldDTO = this.getProjectById(id);
+            ProjectDTO oldDTO = this.getProjectById(projectId);
 
             Date releaseDate = oldDTO.releaseDate();
             if (request.releaseDate() != null) {
@@ -101,10 +96,9 @@ public class ProjectService {
             }
 
             ProjectEntity updatedProjectEntity = ProjectEntity.builder()
-                    .id(id)
+                    .id(projectId)
                     .name(request.name() == null ? oldDTO.name() : request.name())
                     .description(request.description() == null ? oldDTO.description() : request.description())
-                    .namespace(request.namespace() == null ? oldDTO.namespace() : request.namespace())
                     .publisher(request.publisher() == null ? oldDTO.publisher() : request.publisher())
                     .releaseDate(releaseDate)
                     .roles(request.roles() == null ? oldDTO.roles() : request.roles())
@@ -115,78 +109,21 @@ public class ProjectService {
         }
     }
 
-
-    /**
-     * Returns a ProjectDTO based on the provided namespace
-     * @param namespace The namespace of the project.
-     * @return The ProjectEntity's {@code ProjectDTO}.
-     * @throws ProjectNotFoundException When the namespace doesn't belong to any ProjectEntity.
-     */
-    public ProjectDTO getProjectByNamespace(String namespace) throws ProjectNotFoundException {
-        try {
-            String projectId = this.getProjectId(namespace);
-            return this.getProjectById(projectId);
-        } catch (ProjectNotFoundException e) {
-            throw new ProjectNotFoundException(e.getMessage(), e.getCause());
-        }
-    }
-
-
-    /**
-     * Gets the ProjectEntity with the given Namespace.
-     * @param namespace The namespace of the project.
-     * @return The ProjectEntity's {@code ProjectDTO}.
-     * @throws ProjectNotFoundException When the namespace doesn't belong to any ProjectEntity.
-     */
-    public ProjectDTO getSimpleProjectByNamespace(String namespace) throws ProjectNotFoundException {
-        try {
-            String projectId = this.getProjectId(namespace);
-            return this.getProjectById(projectId);
-        } catch (ProjectNotFoundException e) {
-            throw new ProjectNotFoundException(e.getMessage(), e.getCause());
-        }
-    }
-
-
-    /**
-     * Updates the ProjectEntity with the given Namespace
-     * @param namespace The namespace of the project.
-     * @param request The request with the values to be edited.
-     * @throws ProjectNotFoundException When the namespace doesn't belong to any ProjectEntity.
-     */
-    public void updateProjectByNamespace(String namespace, ProjectAddRequest request) throws ProjectNotFoundException {
-        try {
-            String projectId = this.getProjectId(namespace);
-            updateProjectById(projectId, request);
-        } catch (ProjectNotFoundException e) {
-            throw new ProjectNotFoundException(e.getMessage(), e.getCause());
-        }
-    }
-
-
-    /**
-     * Deletes the ProjectEntity with the given Namespace.
-     * @param namespace The namespace of the ProjectEntity.
-     * @throws RuntimeException When the namespace doesn't belong to any ProjectEntity.
-     */
-    public void deleteProjectByNamespace(String namespace) throws ProjectNotFoundException {
-        try {
-            deleteProjectById(getProjectId(namespace));
-        } catch (ProjectNotFoundException e) {
-            throw new ProjectNotFoundException(e.getMessage(), e.getCause());
-        }
-    }
-
-
     private ProjectEntity getProjectFromRequest(ProjectAddRequest request) {
         Date convertedDate = DateHandler.fromString(request.releaseDate());
         return ProjectEntity.builder()
                 .name(request.name())
                 .description(request.description())
-                .namespace(request.namespace())
                 .publisher(request.publisher())
                 .releaseDate(convertedDate)
                 .roles(request.roles())
                 .build();
+    }
+
+    private boolean isUUIDInvalid(String uuid) {
+        if (uuid == null) {
+            return true;
+        }
+        return !UUID_V4_REGEX.matcher(uuid).matches();
     }
 }
